Refactoring WorkloadStore and ServiceStore
This addresses a few things:

* Creates a new `state` package. At the top level, contains `ProxyState` (container for the workloads and service stores), `DemandProxyState` (renamed from `WorkloadInformation`: wraps around `ProxyState` and adds an on-demand API), and `ProxyStateManager` (renamed from `WorkloadManager`)
* Moves `ServiceStore` out of `WorkloadStore`, making it a top-level store within `ProxyState`.
* Replaces `Mutex` with a `RwLock`. The `DemandProxyStore` now has a single `RwLock` for access to the entire `ProxyState`.
* Removes individual `RwLock` around `Workload` and `Service`. Instead, all call sites now just lock (read or write) the entire store.
* Extracted code that converts from XDS and updates the stores and put it into a new structure: `ProxyStateUpdater`. This implements all XDS `Handler` traits.
