+++ src/lib.rs
@@ -15,21 +15,21 @@
 pub mod admin;
 pub mod app;
 pub mod baggage;
+pub mod cert_fetcher;
 pub mod config;
 pub mod identity;
 pub mod metrics;
 pub mod proxy;
 pub mod rbac;
 pub mod readiness;
-pub mod service;
 pub mod signal;
 pub mod socket;
+pub mod state;
 pub mod stats;
 pub mod telemetry;
 pub mod time;
 pub mod tls;
 pub mod version;
-pub mod workload;
 pub mod xds;
 
 pub mod hyper_util;
