+++ src/state/service.rs
@@ -12,18 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::workload::{
+use crate::state::workload::{
     byte_to_ip, network_addr, NamespacedHostname, NetworkAddress, WorkloadError,
 };
+use crate::xds;
 use crate::xds::istio::workload::PortList;
 use bytes::Bytes;
 use std::collections::{HashMap, HashSet};
-use std::sync::{Arc, RwLock};
+use std::ops::Deref;
+use std::sync::Arc;
 use tracing::trace;
-
 use xds::istio::workload::Service as XdsService;
-
-use crate::xds;
 
 #[derive(Debug, Eq, PartialEq, Clone, serde::Serialize, serde::Deserialize)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
@@ -31,8 +30,10 @@ pub struct Service {
     pub name: String,
     pub namespace: String,
     pub hostname: String,
-    pub addresses: Vec<NetworkAddress>,
+    pub vips: Vec<NetworkAddress>,
     pub ports: HashMap<u16, u16>,
+
+    /// Maps workload endpoint addresses to [Endpoint]s.
     #[serde(default)]
     pub endpoints: HashMap<NetworkAddress, Endpoint>,
 }
@@ -49,7 +50,13 @@ impl Service {
 #[derive(Debug, Eq, PartialEq, Clone, serde::Serialize, serde::Deserialize)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
 pub struct Endpoint {
+    /// The service VIP for this endpoint.
+    pub vip: NetworkAddress,
+
+    /// The workload address.
     pub address: NetworkAddress,
+
+    /// The port mapping.
     pub port: HashMap<u16, u16>,
 }
 
@@ -69,12 +76,12 @@ impl TryFrom<&XdsService> for Service {
             name: s.name.to_string(),
             namespace: s.namespace.to_string(),
             hostname: s.hostname.to_string(),
-            addresses: nw_addrs,
+            vips: nw_addrs,
             ports: (&PortList {
                 ports: s.ports.clone(),
             })
                 .into(),
-            endpoints: Default::default(), // intentionally empty; will be populated once inserted into the workload store
+            endpoints: Default::default(), // Will be populated once inserted into the store.
         };
         Ok(svc)
     }
@@ -92,20 +99,19 @@ pub struct ServiceStore {
     workload_to_vips: HashMap<NetworkAddress, HashSet<NetworkAddress>>,
 
     /// Allows for lookup of services by network address, the service's xds secondary key.
-    /// this map stores the same services as `services_by_hostname` so we can mutate the services in both maps
-    /// at once when a service is updated or workload updates affect a service's endpoints.
-    by_vip: HashMap<NetworkAddress, Arc<RwLock<Service>>>,
+    by_vip: HashMap<NetworkAddress, Arc<Service>>,
 
-    /// Allows for lookup of services by namespaced and hostname, the service's xds primary key.
-    /// this map stores the same services as `services_by_ip` so we can mutate the services in both maps
-    /// at once when a service is updated or workload updates affect a service's endpoints.
-    by_host: HashMap<String, Vec<Arc<RwLock<Service>>>>,
+    /// Allows for lookup of services by hostname, and then by namespace. XDS uses a combination
+    /// of hostname and namespace as the primary key. In most cases, there will be a single
+    /// service for a given hostname. However, `ServiceEntry` allows hostnames to be overridden
+    /// on a per-namespace basis.
+    by_host: HashMap<String, Vec<Arc<Service>>>,
 }
 
 impl ServiceStore {
     /// Returns the [Service] matching the given VIP.
-    pub fn get_by_vip(&self, vip: &NetworkAddress) -> Option<Arc<RwLock<Service>>> {
-        self.by_vip.get(vip).cloned()
+    pub fn get_by_vip(&self, vip: &NetworkAddress) -> Option<Service> {
+        self.by_vip.get(vip).map(|s| s.deref().clone())
     }
 
     /// Returns the list of [Service]s matching the given hostname. Istio `ServiceEntry`
@@ -115,8 +121,13 @@ impl ServiceStore {
     /// # Arguments
     ///
     /// * `hostname` - the hostname of the service.
-    pub fn get_by_host(&self, hostname: &String) -> Option<Vec<Arc<RwLock<Service>>>> {
-        self.by_host.get(hostname).cloned()
+    pub fn get_by_host(&self, hostname: &String) -> Option<Vec<Service>> {
+        self.by_host.get(hostname).map(|services| {
+            services
+                .iter()
+                .map(|service| service.deref().clone())
+                .collect()
+        })
     }
 
     /// Returns the [Service] matching the given namespace and hostname. Istio `ServiceEntry`
@@ -126,10 +137,7 @@ impl ServiceStore {
     /// # Arguments
     ///
     /// * `host` - the namespaced hostname.
-    pub fn get_by_namespaced_host(
-        &self,
-        host: &NamespacedHostname,
-    ) -> Option<Arc<RwLock<Service>>> {
+    pub fn get_by_namespaced_host(&self, host: &NamespacedHostname) -> Option<Service> {
         // Get the list of services that match the hostname. Typically there will only be one, but
         // ServiceEntry allows configuring arbitrary hostnames on a per-namespace basis.
         match self.by_host.get(&host.hostname) {
@@ -137,8 +145,8 @@ impl ServiceStore {
             Some(services) => {
                 // Return the service that matches the requested namespace.
                 for service in services {
-                    if service.read().unwrap().namespace == host.namespace {
-                        return Some(service.clone());
+                    if service.namespace == host.namespace {
+                        return Some(service.deref().clone());
                     }
                 }
                 None
@@ -147,31 +155,35 @@ impl ServiceStore {
     }
 
     /// Adds an endpoint for the service VIP.
-    pub(super) fn insert_endpoint(&mut self, vip: &NetworkAddress, ep: Endpoint) {
-        // Insert a mapping from the address to the service VIP.
-        self.workload_to_vips
-            .entry(ep.address.clone())
-            .or_default()
-            .insert(vip.clone());
+    pub fn insert_endpoint(&mut self, ep: Endpoint) {
+        if let Some(svc) = self.by_vip.get(&ep.vip) {
+            // Clone the service and add the endpoint.
+            let mut svc = svc.deref().clone();
+            svc.endpoints.insert(ep.address.clone(), ep);
 
-        if let Some(svc) = self.by_vip.get_mut(vip) {
-            // Add the endpoint to the service.
-            svc.write()
-                .unwrap()
-                .endpoints
-                .insert(ep.address.clone(), ep);
+            // Update the service.
+            self.insert(svc);
         } else {
-            // Can happen due to ordering issues
-            trace!("pod has VIP {vip}, but VIP not found");
+            // We received workload endpoints, but don't have the Service yet.
+            // This can happen due to ordering issues.
+            trace!("pod has VIP {}, but VIP not found", ep.vip);
+
+            // Add a staged entry. This will be added to the service once we receive it.
             self.staged_vips
-                .entry(vip.clone())
+                .entry(ep.vip.clone())
                 .or_default()
-                .insert(ep.address.clone(), ep);
+                .insert(ep.address.clone(), ep.clone());
+
+            // Insert a mapping from the workload address to the VIP.
+            self.workload_to_vips
+                .entry(ep.address.clone())
+                .or_default()
+                .insert(ep.vip.clone());
         }
     }
 
-    /// Removes entries for the given endpoint.
-    pub(super) fn remove_endpoint(&mut self, addr: &NetworkAddress) {
+    /// Removes entries for the given endpoint address.
+    pub fn remove_endpoint(&mut self, addr: &NetworkAddress) {
         // Remove the endpoint from the VIP map.
         let Some(prev_vips) = self.workload_to_vips.remove(addr) else {
             return;
@@ -188,59 +200,66 @@ impl ServiceStore {
             }
 
             // Remove the endpoint from the service.
-            if let Some(wls) = self.by_vip.get_mut(vip) {
-                wls.write().unwrap().endpoints.remove(addr);
+            if let Some(service) = self.by_vip.get(vip) {
+                // Clone the service and remove the endpoint.
+                let mut service = service.deref().clone();
+                service.endpoints.remove(addr);
+
+                // Update the service.
+                self.insert(service);
             }
         }
     }
 
     /// Adds the given service.
-    pub(super) fn insert(&mut self, mut service: Service) {
+    pub fn insert(&mut self, mut service: Service) {
         // First mutate the service and add all missing endpoints
-        for addr in service.addresses.as_slice() {
+        for vip in &service.vips {
             // Due to ordering issues, we may have gotten workloads with VIPs before we got the service
             // we should add those workloads to the vips map now
-            if let Some(wips_to_endpoints) = self.staged_vips.remove(addr) {
-                for (wip, ep) in wips_to_endpoints {
+            if let Some(endpoints) = self.staged_vips.remove(vip) {
+                for (wip, ep) in endpoints {
                     service.endpoints.insert(wip.clone(), ep);
-                }
-            }
-
-            // If svc already exists, add old endpoints to new svc
-            if let Some(prev) = self.by_vip.get_mut(addr) {
-                let prev = prev.read().unwrap();
-                for (wip, ep) in prev.endpoints.iter() {
-                    service.endpoints.insert(wip.clone(), ep.clone());
                 }
             }
         }
 
         // If we're replacing an existing service, remove the old one from all data structures.
-        self.remove(&service.namespaced_hostname());
+        let _ = self.remove(&service.namespaced_hostname());
 
         // Save values used for the indexes.
-        let addresses = service.addresses.clone();
+        let vips = service.vips.clone();
         let hostname = service.hostname.clone();
 
         // Create the Arc.
-        let service = Arc::new(RwLock::new(service));
+        let service = Arc::new(service);
 
-        // Add the service to the data structures.
-        for addr in &addresses {
-            self.by_vip.insert(addr.clone(), service.clone());
+        // Map the vips to the service.
+        for vip in &vips {
+            self.by_vip.insert(vip.clone(), service.clone());
         }
+
+        // Map the hostname to the service.
         match self.by_host.get_mut(&hostname) {
             None => {
-                let _ = self.by_host.insert(hostname.clone(), vec![service]);
+                let _ = self.by_host.insert(hostname.clone(), vec![service.clone()]);
             }
             Some(services) => {
-                services.push(service);
+                services.push(service.clone());
             }
+        }
+
+        // Map the workload address to the endpoint.
+        for (_, ep) in service.endpoints.iter() {
+            self.workload_to_vips
+                .entry(ep.address.clone())
+                .or_default()
+                .insert(ep.vip.clone());
         }
     }
 
     /// Removes the service for the given host and namespace.
-    pub(super) fn remove(&mut self, namespaced_host: &NamespacedHostname) -> bool {
+    pub fn remove(&mut self, namespaced_host: &NamespacedHostname) -> Option<Service> {
         // Remove the previous service from the by_host map.
         let mut remove_hostname = false;
         let prev = {
@@ -252,9 +271,7 @@ impl ServiceStore {
 
                     let mut i = 0;
                     while i < services.len() {
-                        if services[i].read().unwrap().namespace.as_str()
-                            == namespaced_host.namespace
-                        {
+                        if services[i].namespace.as_str() == namespaced_host.namespace {
                             found = Some(services.remove(i));
 
                             // If the array is empty, also remove the Vec.
@@ -269,41 +286,37 @@ impl ServiceStore {
             }
         };
 
-        // Lock the service for read.
-        let prev = match &prev {
-            None => {
-                // Remove unsuccessful - not found.
-                return false;
-            }
-            Some(prev) => prev.read().unwrap(),
-        };
-
-        // If the Vec for the hostname is empty now, remove it.
-        if remove_hostname {
-            self.by_host.remove(&prev.hostname);
-        }
-
-        // Remove the entries for the previous service IPs.
-        prev.addresses.iter().for_each(|addr| {
-            self.by_vip.remove(addr);
-        });
-
-        // Remove mapping from workload to the VIPs for this service.
-        for (ep_ip, _) in prev.endpoints.iter() {
-            self.workload_to_vips.remove(ep_ip);
-            for network_addr in prev.addresses.iter() {
-                self.staged_vips
-                    .entry(network_addr.clone())
-                    .or_default()
-                    .remove(ep_ip);
-                if self.staged_vips[network_addr].is_empty() {
-                    self.staged_vips.remove(network_addr);
+        match prev {
+            None => None,
+            Some(prev) => {
+                // If the Vec for the hostname is empty now, remove it.
+                if remove_hostname {
+                    self.by_host.remove(&prev.hostname);
                 }
+
+                // Remove the entries for the previous service IPs.
+                prev.vips.iter().for_each(|addr| {
+                    self.by_vip.remove(addr);
+                });
+
+                // Remove mapping from workload to the VIPs for this service.
+                for (ep_ip, _) in prev.endpoints.iter() {
+                    self.workload_to_vips.remove(ep_ip);
+                    for network_addr in prev.vips.iter() {
+                        self.staged_vips
+                            .entry(network_addr.clone())
+                            .or_default()
+                            .remove(ep_ip);
+                        if self.staged_vips[network_addr].is_empty() {
+                            self.staged_vips.remove(network_addr);
+                        }
+                    }
+                }
+
+                // Remove successful.
+                Some(prev.deref().clone())
             }
         }
-
-        // Remove successful.
-        true
     }
 
     #[cfg(test)]
